requirements.txt
requests
pyyaml
confluent_kafka
flask
gunicorn
cachetools